Consolidate setup scripts and move MathTypes
