+++ docs/BUILD.md
@@ -5,14 +5,10 @@ macOS the provided scripts will create an Xcode project configured for
 both macOS and iOS targets.
 
 ```bash
-./scripts/xcodegen.sh
+./scripts/build_setup.sh --open
 ```
 
-This script generates a `build/` directory and opens `FinalStorm.xcodeproj`.
-For manual CMake invocation run:
-
-```bash
-./scripts/gen_xcode.sh
-```
-
-After configuring you can build either from Xcode or via `cmake --build`.
+This script generates a `build/` directory and optionally opens
+`FinalStorm.xcodeproj` when the `--open` flag is supplied. Pass `--ios`
+to configure for the iOS SDK. After configuring you can build either
+from Xcode or via `cmake --build`.
