+++ src/Core/Input/InputTypes.h
@@ -3,7 +3,7 @@
 // Common types for input handling
 
 #pragma once
-#include "Core/MathTypes.h"
+#include "Core/Math/MathTypes.h"
 
 namespace FinalStorm {
