+++ src/Rendering/Light.h
@@ -1,7 +1,7 @@
 // Rendering/Light.h
 #pragma once
 
-#include "Core/MathTypes.h"
+#include "Core/Math/MathTypes.h"
 #include "Core/Transform.h"
 
 namespace FinalStorm {
