+++ src/Rendering/Material.h
@@ -1,7 +1,7 @@
 // Rendering/Material.h
 #pragma once
 
-#include "Core/MathTypes.h"
+#include "Core/Math/MathTypes.h"
 #include "Core/ResourceManager.h"
 #include <string>
 #include <unordered_map>
