+++ src/UI/UI3DPanel.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <memory>
-#include "Core/MathTypes.h"
+#include "Core/Math/MathTypes.h"
 
 namespace FinalStorm {
